add confirm dialog

+++ src/modules/profile/pages/address/addressDetails.vue
@@ -54,10 +54,15 @@ export default {
       this.$router.push({ name: 'AddressList' })
     },
     async deleteHandler() {
-      this.loading = true
-      await addressService.deleteById(this.id)
-      this.loading = false
-      this.$router.go(-1)
+      const res = await this.$confirm(
+        'Вы действительно хотите удалить запись? '
+      )
+      if (res) {
+        this.loading = true
+        await addressService.deleteById(this.id)
+        this.loading = false
+        this.$router.go(-1)
+      }
     },
   },
 }
