+++ src/modules/profile/pages/truck/truckDetails.vue
@@ -59,10 +59,15 @@ export default {
       this.$router.push({ name: 'TruckList' })
     },
     async deleteHandler() {
-      this.loading = true
-      await service.deleteById(this.id)
-      this.loading = false
-      this.$router.go(-1)
+      const res = await this.$confirm(
+        'Вы действительно хотите удалить запись? '
+      )
+      if (res) {
+        this.loading = true
+        await service.deleteById(this.id)
+        this.loading = false
+        this.$router.go(-1)
+      }
     },
   },
 }
